Moved the components to a dedicated directory
