+++ .eslintrc.js
@@ -21,6 +21,7 @@ module.exports = {
   rules: {
     'class-methods-use-this': 'off',
     'import/no-extraneous-dependencies': 'off',
+    'import/prefer-default-export': 'off',
     'no-restricted-syntax': 'off',
     'no-unused-vars': 'off',
     semi: ['error', 'never'],
