+++ lib/SmartDataTable.js
@@ -2,11 +2,11 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 // Import components
-import ErrorBoundary from './ErrorBoundary'
-import TableCell from './TableCell'
-import Toggles from './Toggles'
-import Paginator from './Paginator'
-import withPagination from './with-pagination'
+import ErrorBoundary from './components/ErrorBoundary'
+import TableCell from './components/TableCell'
+import Toggles from './components/Toggles'
+import Paginator from './components/Paginator'
+import withPagination from './components/helpers/with-pagination'
 // Import functions
 import {
   debugPrint,
@@ -22,7 +22,7 @@ import {
 // Import styles
 import './css/basic.css'
 
-class SmartDataTablePlain extends React.Component {
+class SmartDataTable extends React.Component {
   constructor(props) {
     super(props)
 
@@ -343,7 +343,7 @@ longer valid. Consider replacing it with '${newName}'
 }
 
 // Defines the type of data expected in each passed prop
-SmartDataTablePlain.propTypes = {
+SmartDataTable.propTypes = {
   data: PropTypes.oneOfType([PropTypes.string, PropTypes.array]).isRequired,
   dataKey: PropTypes.string,
   columns: PropTypes.array,
@@ -369,7 +369,7 @@ SmartDataTablePlain.propTypes = {
 }
 
 // Defines the default values for not passing a certain prop
-SmartDataTablePlain.defaultProps = {
+SmartDataTable.defaultProps = {
   dataKey: 'data',
   columns: [],
   name: 'reactsmartdatatable',
@@ -391,12 +391,5 @@ SmartDataTablePlain.defaultProps = {
   orderedHeaders: [],
   hideUnordered: false,
 }
-
-// Wrap the component with an Error Boundary
-const SmartDataTable = (props) => (
-  <ErrorBoundary>
-    <SmartDataTablePlain {...props} />
-  </ErrorBoundary>
-)
 
 export default SmartDataTable
