+++ lib/components/Paginator.js
@@ -1,8 +1,8 @@
 // Import modules
 import React from 'react'
 import PropTypes from 'prop-types'
-import { generatePagination, isUndefined } from './helpers/functions'
-import './css/paginator.css'
+import { generatePagination, isUndefined } from '../helpers/functions'
+import '../css/paginator.css'
 
 const PaginatorItem = ({ active, value, text, onPageChange }) => (
   <button
