+++ lib/components/TableCell.js
@@ -9,7 +9,7 @@ import {
   isEmpty,
   head,
   memoizedHighlightValueParts as highlightValueParts,
-} from './helpers/functions'
+} from '../helpers/functions'
 
 class TableCell extends React.Component {
   getRenderValue() {
