+++ lib/components/Toggles.js
@@ -1,7 +1,7 @@
 // Import modules
 import React from 'react'
 import PropTypes from 'prop-types'
-import './css/toggles.css'
+import '../css/toggles.css'
 
 class Toggles extends React.Component {
   constructor(props) {
