+++ lib/components/__tests__/Paginator.js
