+++ lib/components/helpers/with-pagination.js
